// src/types/api.ts
export interface paths {
// ...
    "/api/rest/projects/{project_id}": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /**
                 * @description The project id
                 * @example {{projectId}}
                 */
                project_id: string;
            };
            cookie?: never;
        };
        /** Get a project */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    /**
                     * @description The project id
                     * @example {{projectId}}
                     */
                    project_id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
// ...
                    content: {
// ...
                        "application/json": {
                            projects?: {
                                id?: number;
                                name?: string;
                                status?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                description?: string;
                                enabled?: boolean;
                                view_state?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                access_level?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                custom_fields?: {
                                    id?: number;
                                    name?: string;
                                    type?: string;
                                    default_value?: string;
                                    possible_values?: string;
                                    valid_regexp?: string;
                                    length_min?: number;
                                    length_max?: number;
                                    access_level_r?: {
                                        id?: number;
                                        name?: string;
                                        label?: string;
                                    };
                                    access_level_rw?: {
                                        id?: number;
                                        name?: string;
                                        label?: string;
                                    };
                                    display_report?: boolean;
                                    display_update?: boolean;
                                    display_resolved?: boolean;
                                    display_closed?: boolean;
                                    require_report?: boolean;
                                    require_update?: boolean;
                                    require_resolved?: boolean;
                                    require_closed?: boolean;
                                }[];
                                versions?: {
                                    id?: number;
                                    name?: string;
                                    description?: string;
                                    released?: boolean;
                                    obsolete?: boolean;
                                    /** Format: date-time */
                                    timestamp?: string;
                                }[];
                                subProjects?: {
                                    id?: number;
                                    name?: string;
                                }[];
                            }[];
                        };
                    };
                };
            };
        };
// ...
    };
// ...
}